Adds the ability to require()
This makes it possible to import both other scripts, as well as single
functions (both named and anonymous) into the worker.

A path to a script given to require() is relative to the worker. This
means we have to use eval.js in order to have a fixed path to use as a
reference.